+++ lib/global/theme/app_themes.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 
 enum AppTheme { Blue, Ogange, Green }
 
+
+
 final appThemeData = {
   AppTheme.Blue: ThemeData(
       textTheme: TextTheme(bodyText1: TextStyle(color: Colors.white)),
