+++ lib/global/theme/bloc/theme_event.dart
@@ -1,9 +1,7 @@
 part of 'theme_bloc.dart';
 
-abstract class ThemeEvent  {
+abstract class ThemeEvent {
   const ThemeEvent();
-
-  
 }
 
 class ThemeChanged extends ThemeEvent {
