+++ lib/global/theme/bloc/theme_state.dart
@@ -1,9 +1,10 @@
 part of 'theme_bloc.dart';
 
-class ThemeState  {
+class ThemeState {
   final ThemeData? themeData;
+  
 
   const ThemeState(this.themeData);
 
-  
+  const ThemeState.dark(this.themeData);
 }
