+++ lib/screens/login.dart
@@ -24,85 +24,129 @@ class LoginState extends State<Login> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-        backgroundColor: Theme.of(context).scaffoldBackgroundColor,
-        appBar: AppBar(
-          iconTheme: IconThemeData(color: Theme.of(context).backgroundColor),
-          title: Text(
-            'Welcome',
-            style: TextStyle(color: Theme.of(context).backgroundColor),
+    return SafeArea(
+        child: Container(
+      color: Theme.of(context).colorScheme.primary,
+      child: Stack(
+        alignment: AlignmentDirectional.topCenter,
+        children: [
+          Container(
+            decoration: BoxDecoration(
+                color: Colors.black,
+                border: Border.all(
+                    width: 0,
+                    color: Theme.of(context)
+                        .colorScheme
+                        .primary
+                        .withOpacity(0.7))),
+            height: MediaQuery.of(context).size.height * 0.3,
           ),
-          backgroundColor: Theme.of(context).scaffoldBackgroundColor,
-          actions: [
-            Hero(tag: 'icon', child: getIcon(context, size: 100.0)),
-          ],
-        ),
-        body: Padding(
-            padding: EdgeInsets.all(16),
-            child: Center(
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Container(
-                    height: 236,
-                    padding: EdgeInsets.all(15),
-                    decoration: BoxDecoration(
-                      color: Theme.of(context).backgroundColor,
-                      borderRadius: BorderRadius.all(Radius.circular(20)),
-                    ),
-                    child: Column(
-                      children: [
-                        Padding(
-                            padding: EdgeInsets.only(bottom: 16),
-                            child: TextFormField(
-                              validator: (value) =>
-                                  value!.isEmpty || value.contains('@')
-                                      ? 'Email cannot be blank'
-                                      : null,
-                              keyboardType: TextInputType.emailAddress,
-                              controller: emailController,
-                              decoration: InputDecoration(
-                                  focusColor:
-                                      Theme.of(context).colorScheme.primary,
-                                  border: OutlineInputBorder(),
-                                  labelText: 'Email'),
-                            )),
-                        Padding(
-                            padding: EdgeInsets.only(bottom: 16),
-                            child: TextFormField(
-                              validator: (value) =>
-                                  value!.isEmpty || value.length < 6
-                                      ? 'Minimum 6 symbols'
-                                      : null,
-                              controller: passwordController,
-                              obscureText: true,
-                              decoration: InputDecoration(
-                                  hintText: 'Minimum 6 symbols',
-                                  border: OutlineInputBorder(),
-                                  labelText: 'Password'),
-                            )),
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [
-                            RegisterButton(
-                                auth: _auth,
-                                emailController: emailController,
-                                passwordController: passwordController),
-                            SingInButton(
-                                auth: _auth,
-                                emailController: emailController,
-                                passwordController: passwordController),
-                          ],
-                        ),
-                      ],
-                    ),
-                  ),
-                  SizedBox(
-                    height: 20,
-                  ),
-                ],
-              ),
-            )));
+          Container(
+            width: double.infinity,
+            decoration: BoxDecoration(
+                gradient: LinearGradient(colors: [
+                  Theme.of(context).colorScheme.primary.withOpacity(0.7),
+                  Theme.of(context).colorScheme.primary,
+                ], begin: Alignment.topRight, end: Alignment.bottomLeft),
+                border: Border.all(
+                  width: 0,
+                  color: Theme.of(context).colorScheme.primary.withOpacity(0.7),
+                ),
+                borderRadius:
+                    BorderRadius.only(bottomRight: Radius.circular(75))),
+            height: MediaQuery.of(context).size.height * 0.3,
+          ),
+          Container(
+              margin: EdgeInsets.only(
+                  top: MediaQuery.of(context).size.height * 0.2,
+                  left: MediaQuery.of(context).size.width * 0.7),
+              child: getIcon(context, size: 60.0)),
+          Container(
+            child: Material(
+              color: Colors.black,
+              child: LoginForm(
+                  emailController: emailController,
+                  passwordController: passwordController,
+                  auth: _auth),
+            ),
+            margin:
+                EdgeInsets.only(top: MediaQuery.of(context).size.height * 0.3),
+            decoration: BoxDecoration(
+                border: Border.all(
+                  width: 0,
+                ),
+                color: Colors.black,
+                borderRadius: BorderRadius.only(topLeft: Radius.circular(75))),
+            width: double.infinity,
+            height: MediaQuery.of(context).size.height * 0.7,
+          )
+        ],
+      ),
+    ));
+  }
+}
+
+class LoginForm extends StatelessWidget {
+  const LoginForm({
+    Key? key,
+    required this.emailController,
+    required this.passwordController,
+    required FirebaseAuth auth,
+  })  : _auth = auth,
+        super(key: key);
+
+  final TextEditingController emailController;
+  final TextEditingController passwordController;
+  final FirebaseAuth _auth;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: EdgeInsets.all(20),
+      child: Column(
+        children: [
+          Padding(
+              padding: EdgeInsets.only(bottom: 16),
+              child: TextFormField(
+                validator: (value) => value!.isEmpty || value.contains('@')
+                    ? 'Email cannot be blank'
+                    : null,
+                keyboardType: TextInputType.emailAddress,
+                controller: emailController,
+                decoration: InputDecoration(
+                    focusColor: Theme.of(context).colorScheme.primary,
+                    border: OutlineInputBorder(),
+                    labelText: 'Email'),
+              )),
+          Padding(
+              padding: EdgeInsets.only(bottom: 16),
+              child: TextFormField(
+                validator: (value) => value!.isEmpty || value.length < 6
+                    ? 'Minimum 6 symbols'
+                    : null,
+                controller: passwordController,
+                obscureText: true,
+                decoration: InputDecoration(
+                    hintText: 'Minimum 6 symbols',
+                    border: OutlineInputBorder(),
+                    labelText: 'Password'),
+              )),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              RegisterButton(
+                  auth: _auth,
+                  emailController: emailController,
+                  passwordController: passwordController),
+              SingInButton(
+                  auth: _auth,
+                  emailController: emailController,
+                  passwordController: passwordController),
+            ],
+          ),
+        ],
+      ),
+    );
   }
 }
